Add styles jss files
