+++ freedom-for-data-client/src/shared/SidebarStyles.js
@@ -1,0 +1,10 @@
+const styles = {
+  list: {
+    width: 250,
+  },
+  fullList: {
+    width: 'auto',
+  },
+};
+
+export default styles;
